cs50-week0-ScratchProject
In this project I have made a simple game of bouncing ball that loses lives if it touches the bottom edge. You must use the left and right keys to move the paddle in order to avoid losing lives. Be careful you only have 3 lives and the difficulty of the game increases as you make progress.